Simple overflow checking on GetBytes and GetChars
With the methods that return the number of bytes that are read, it should be expected that they should safeguard against reading over the buffer length. Before this patch, if you stuck in a length that was larger than the rest of the buffer size, the Array.Copy would throw an argument exception. Now, after this patch it behaves like Stream.Read.